% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_recode.R
\name{ind_recode}
\alias{ind_recode}
\title{ind_recode}
\usage{
ind_recode(
  var,
  var_prefix = "ind_",
  negator = "not",
  cat_true = NULL,
  cat_false = NULL,
  rev = FALSE
)
}
\arguments{
\item{var}{the name of an indicator variable}

\item{var_prefix}{a character string that will be ignored when creating the categorical variable}

\item{negator}{a character string used to create cat_false when cat_false is NULL, default is 'not'}

\item{cat_true}{a character string string to be used place of  T/1/"Yes" for the categorical variable output, if NULL the category is automatically generated from the variable name}

\item{cat_false}{a character string string to be used place of  F/0/"No" for the categorical variable output, if NULL the category is automatically generated from the cat true and the negator}

\item{rev}{logical indicating if the order should be reversed from the F/T ordering of the indicator source variable, default is FALSE}
}
\description{
ind_recode
}
\examples{
library(tibble)
library(dplyr)
tibble(ind_grad = c(0,0,1,1,1 ,0 ,0)) \%>\%
  mutate(cat_grad  = ind_recode(ind_grad))

tibble(ind_grad = c(TRUE,TRUE,FALSE)) \%>\%
  mutate(cat_grad  = ind_recode(ind_grad))

tibble(ind_grad = c("Y", "N")) \%>\%
  mutate(cat_grad  = ind_recode(ind_grad))

tibble(ind_grad = c("y", "n")) \%>\%
  mutate(cat_grad  = ind_recode(ind_grad))

tibble(ind_grad = c("yes", "no")) \%>\%
  mutate(cat_grad  = ind_recode(ind_grad))
}
